Implement preprocess_input method for head_pose
Along with preprocess_input method update for head_pose module, the predict method for FaceDetection is also updated to return also a flag if a face is detected or not.
app.py is modified to proceed to the next models if a face is detected or skip to next frame if not.